+++ src/keyrecon/__init__.py
@@ -9,4 +9,12 @@ __all__ = [
     "KeyRecon",
     "AdaptiveCKSWeights",
     "EN_REFERENCE_WEIGHTS",
+    "KoreanReference",
+    "load_korean_reference",
+    "KoreanAdaptiveCKS",
+    "KoreanRunResult",
+    "run_korean_adaptive",
 ]
+
+from .ko_reference import KoreanReference, load_korean_reference
+from .ko_runtime import KoreanAdaptiveCKS, KoreanRunResult, run_korean_adaptive